python-service/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
redis==5.0.7
setuptools==80.9.0
opentelemetry-distro==0.46b0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-instrumentation-psycopg2==0.46b0
opentelemetry-instrumentation-redis==0.46b0
